fix scroll down caching bug
If cursor wasn't at the bottom of screen when terminal got down
to rendering lines that needed to be off-screen (scroll the screen
down), scroll_down calls in the final loop wouldn't move the screen
down. Caching causes this because the bottom line hasn't been moved to,
since no writing there was required because it matched what was there
already.