Changed curtsies.formatstringarray.FSArray references to curtsies.FSArray
Not sure if this is the right thing to do, but it seems to me that users
will access FSArrays directly via curtsies.FSArray so this is less confusing.
In any case, the documentation for FSArray was being repeated at the bottom
so we should probably pick one format and stick with it?