Test if Input has an orig_sigint_handler before attempting to use it
I don't know what's going on or why this is occuring, but I'm getting
a exception due to this being none while trying to embed bpython into
the lldb debugger. For whatever reason the sigint handler is None
during __enter__ and thus this throws.

Simply test if None before attempting to use.